# src/lib/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpSolid Public API",
    "version": "1.0.0",
    "description": "Versioned public API for OpSolid Digital Business Cards. Bearer-only JWT auth on `/auth/*` (after login) and `/cards/*`. Public reads at `/public/cards/*`. Stable contract — fields will not be renamed or removed without a v2 bump.",
    "contact": { "email": "support@opsolid.de" }
  },
  "servers": [
    { "url": "https://opsolid.de/api/v1", "description": "Production" },
    { "url": "https://card.opsolid.de/api/v1", "description": "Production (card host)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256-signed access token returned by /auth/login or /auth/refresh. 15 minute TTL — refresh proactively from the client."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "invalid_credentials" },
              "message": { "type": "string", "example": "Invalid email or password." },
              "details": {}
            }
          }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "email"],
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": ["string", "null"] },
          "locale": { "type": ["string", "null"], "enum": ["de", "en", "tr", null] },
          "emailVerifiedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "AuthTokenPair": {
        "type": "object",
        "required": ["accessToken", "refreshToken", "sessionExpiresAt", "user"],
        "properties": {
          "accessToken": { "type": "string", "description": "Short-lived (15 min) HS256 JWT." },
          "refreshToken": { "type": "string", "description": "Opaque rotation token. Single-use; replaced on every /auth/refresh." },
          "sessionExpiresAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid bearer token.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": { "schema": { "type": "string" } }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "BadRequest": {
        "description": "Invalid payload.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Sign in with email + password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthTokenPair" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Rotate refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotated.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthTokenPair" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  }
}